+++ src/core/dwn-error.ts
@@ -52,7 +52,6 @@ export enum DwnErrorCode {
   RecordsProtocolContextDerivationSchemeMissingContextId = 'RecordsProtocolContextDerivationSchemeMissingContextId',
   RecordsProtocolPathDerivationSchemeMissingProtocol = 'RecordsProtocolPathDerivationSchemeMissingProtocol',
   RecordsSchemasDerivationSchemeMissingSchema = 'RecordsSchemasDerivationSchemeMissingSchema',
-  RecordsReadMissingDescriptorProperties = 'RecordsReadMissingDescriptorProperties',
   RecordsReadMissingCreateProperties = 'RecordsReadMissingCreateProperties',
   RecordsWriteGetEntryIdUndefinedAuthor = 'RecordsWriteGetEntryIdUndefinedAuthor',
   RecordsWriteDataCidMismatch = 'RecordsWriteDataCidMismatch',
